requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
anyio==4.4.0
    # via
    #   gql
    #   httpx
    #   jupyter-server
argon2-cffi==23.1.0
    # via jupyter-server
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.3.0
    # via isoduration
asttokens==2.4.1
    # via stack-data
async-lru==2.0.4
    # via jupyterlab
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
babel==2.15.0
    # via jupyterlab-server
backoff==2.2.1
    # via gql
beautifulsoup4==4.12.3
    # via nbconvert
bleach==6.1.0
    # via nbconvert
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.0
    # via argon2-cffi-bindings
charset-normalizer==3.3.2
    # via requests
comm==0.2.2
    # via ipykernel
cytoolz==0.12.3
    # via eth-utils
debugpy==1.8.5
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
eth-abi==4.2.1
    # via python-cartesi
eth-hash==0.7.0
    # via eth-utils
eth-typing==4.4.0
    # via
    #   eth-abi
    #   eth-utils
eth-utils==4.1.1
    # via eth-abi
executing==2.0.1
    # via stack-data
fastjsonschema==2.20.0
    # via nbformat
fqdn==1.5.1
    # via jsonschema
gql[requests]==3.5.0
    # via -r requirements.in
graphql-core==3.2.3
    # via gql
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via jupyterlab
idna==3.7
    # via
    #   anyio
    #   httpx
    #   jsonschema
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
ipykernel==6.29.5
    # via jupyterlab
ipython==8.26.0
    # via ipykernel
isoduration==20.11.0
    # via jsonschema
jedi==0.19.1
    # via ipython
jinja2==3.1.4
    # via
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
json5==0.9.25
    # via jupyterlab-server
jsonpointer==3.0.0
    # via jsonschema
jsonschema[format-nongpl]==4.23.0
    # via
    #   jupyter-events
    #   jupyterlab-server
    #   nbformat
jsonschema-specifications==2023.12.1
    # via jsonschema
jupyter-client==8.6.2
    # via
    #   ipykernel
    #   jupyter-server
    #   nbclient
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   nbclient
    #   nbconvert
    #   nbformat
jupyter-events==0.10.0
    # via jupyter-server
jupyter-lsp==2.2.5
    # via jupyterlab
jupyter-server==2.14.2
    # via
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   notebook-shim
jupyter-server-terminals==0.5.3
    # via jupyter-server
jupyterlab==4.2.4
    # via -r requirements.in
jupyterlab-pygments==0.3.0
    # via nbconvert
jupyterlab-server==2.27.3
    # via jupyterlab
markupsafe==2.1.5
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mistune==3.0.2
    # via nbconvert
multidict==6.0.5
    # via yarl
nbclient==0.10.0
    # via nbconvert
nbconvert==7.16.4
    # via jupyter-server
nbformat==5.10.4
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
nest-asyncio==1.6.0
    # via ipykernel
notebook-shim==0.2.4
    # via jupyterlab
numpy==2.0.1
    # via pandas
overrides==7.7.0
    # via jupyter-server
packaging==24.1
    # via
    #   ipykernel
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
    #   pytest
pandas==2.2.2
    # via -r requirements.in
pandocfilters==1.5.1
    # via nbconvert
parsimonious==0.9.0
    # via eth-abi
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
platformdirs==4.2.2
    # via jupyter-core
pluggy==1.5.0
    # via pytest
prometheus-client==0.20.0
    # via jupyter-server
prompt-toolkit==3.0.47
    # via ipython
psutil==6.0.0
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.3
    # via stack-data
pycparser==2.22
    # via cffi
pycryptodome==3.19.1
    # via python-cartesi
pydantic==1.10.17
    # via python-cartesi
pygments==2.18.0
    # via
    #   ipython
    #   nbconvert
pytest==7.4.4
    # via python-cartesi
python-cartesi==0.1.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   jupyter-client
    #   pandas
python-json-logger==2.0.7
    # via jupyter-events
pytz==2024.1
    # via pandas
pyyaml==6.0.2
    # via jupyter-events
pyzmq==26.1.0
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   jupyter-events
regex==2024.7.24
    # via parsimonious
requests==2.32.3
    # via
    #   -r requirements.in
    #   gql
    #   jupyterlab-server
    #   python-cartesi
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via gql
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
send2trash==1.8.3
    # via jupyter-server
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
soupsieve==2.5
    # via beautifulsoup4
stack-data==0.6.3
    # via ipython
terminado==0.18.1
    # via
    #   jupyter-server
    #   jupyter-server-terminals
tinycss2==1.3.0
    # via nbconvert
toolz==0.12.1
    # via cytoolz
tornado==6.4.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   terminado
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
types-python-dateutil==2.9.0.20240316
    # via arrow
typing-extensions==4.12.2
    # via
    #   eth-typing
    #   pydantic
tzdata==2024.1
    # via pandas
uri-template==1.3.0
    # via jsonschema
urllib3==2.2.2
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
webcolors==24.6.0
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.8.0
    # via jupyter-server
yarl==1.9.4
    # via gql

# The following packages are considered to be unsafe in a requirements file:
# setuptools
